Reorganize auth helpers
