+++ src/Gameboard.Api/Structure/Auth/ApiKeyAuthenticationService.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+
+namespace Gameboard.Api.Auth;
+
+internal class ApiKeyAuthenticationService : IApiKeyAuthenticationService
+{
+    public Task<string> ResolveApiKey(string key)
+    {
+        throw new System.NotImplementedException();
+    }
+}
