+++ src/Gameboard.Api/Structure/Auth/AuthenticationStartupExtensions.cs
@@ -1,9 +1,9 @@
 // Copyright 2021 Carnegie Mellon University. All Rights Reserved.
 // Released under a MIT (SEI)-style license. See LICENSE.md in the project root for license information.
 
-using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
 using Gameboard.Api;
+using Gameboard.Api.Auth;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Http;
