+++ src/Gameboard.Api/Structure/Auth/ClaimsPrincipalExtensions.cs
